refactor: OrderLineItems Wrapper class 생성
