+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,15 +1,11 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.order.exception.EmptyOrderLineItemsException;
 import kitchenpos.order.exception.InvalidChangeOrderStatusException;
-import org.springframework.util.CollectionUtils;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 @Entity
 @Table(name = "orders")
@@ -28,9 +24,8 @@ public class Order {
     @Column(nullable = false)
     private LocalDateTime orderedTime;
 
-    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
-    @JoinColumn(name = "order_id")
-    private List<OrderLineItem> orderLineItems = new ArrayList<>();
+    @Embedded
+    private OrderLineItems orderLineItems;
 
     public Order() {
     }
@@ -38,16 +33,9 @@ public class Order {
     private Order(Long id, Long orderTableId, OrderStatus orderStatus, List<OrderLineItem> orderLineItems) {
         this.id = id;
         this.orderTableId = orderTableId;
-        this.orderLineItems = orderLineItems;
+        this.orderLineItems = new OrderLineItems(orderLineItems);
         this.orderStatus = OrderStatus.COOKING;
         this.orderedTime = LocalDateTime.now();
-        validate();
-    }
-
-    private void validate() {
-        if (CollectionUtils.isEmpty(orderLineItems)) {
-            throw new EmptyOrderLineItemsException("주문 항목이 없어 등록할 수 없습니다.");
-        }
     }
 
     public static Order of(Long id, Long orderTableId, OrderStatus orderStatus, List<OrderLineItem> orderLineItems) {
@@ -55,9 +43,7 @@ public class Order {
     }
 
     public void addOrderIdToOrderLineItems() {
-        orderLineItems = orderLineItems.stream()
-                .map(orderLineItem -> orderLineItem.addOrderId(id))
-                .collect(Collectors.toList());
+        orderLineItems.addOrderIdToOrderLineItems(id);
     }
 
     public void checkOrderStatusIsCompletion() {
@@ -87,12 +73,10 @@ public class Order {
     }
 
     public List<OrderLineItem> getOrderLineItems() {
-        return orderLineItems;
+        return orderLineItems.getOrderLineItems();
     }
 
     public List<Long> getMenuIds() {
-        return orderLineItems.stream()
-                .map(OrderLineItem::getMenuId)
-                .collect(Collectors.toList());
+        return orderLineItems.getMenuIds();
     }
 }
